+++ pages/boards/[id]/edit/index.tsx
@@ -12,6 +12,7 @@ import useEditStore from "@/store/edit";
 import Modal from "@/components/modal/modal";
 import ModalCheckIt from "@/components/modal/modalCheckIt";
 import useSideStore from "@/store/side";
+import InviteModal from "@/components/inviteModal";
 
 export default function Edit() {
   const { setSide } = useSideStore();
@@ -21,11 +22,7 @@ export default function Edit() {
   const [emailListData, setEmailListData] = useState<any>();
   const [dashboardData, setDashboardData] = useState<any>();
   const [inviteEmailInput, setInviteEmailInput] = useState<string>("");
-  const [canIInvite, setCanIInvite] = useState<boolean>(false);
-  const [errorModal, setErrorModal] = useState<boolean>(false);
-  const [errorModal2, setErrorModal2] = useState<boolean>(false);
-  const { inputState, colorState, inviteModalState, setInviteModalState } =
-    useEditStore();
+  const { inputState, colorState, inviteModalState } = useEditStore();
 
   const router = useRouter();
   const { id } = router.query;
@@ -69,20 +66,6 @@ export default function Edit() {
     }
   };
 
-  const postData = async (link: string, data: any) => {
-    try {
-      const response = await axios.post(link, data);
-    } catch (e: any) {
-      if (e.response.status === 409) {
-        await setErrorModal(true);
-        setInviteModalState(false);
-      } else if (e.response.status === 404) {
-        await setErrorModal2(true);
-        setInviteModalState(false);
-      }
-    }
-  };
-
   const handleEditDashboard = async () => {
     await putData(`dashboards/${id}`, {
       title: inputState,
@@ -117,31 +100,6 @@ export default function Edit() {
     }
   };
 
-  const handleSetInviteModalStateFalse = () => {
-    setInviteModalState(false);
-    setErrorModal(false);
-    setErrorModal2(false);
-  };
-
-  const handleInviteEmail = async () => {
-    const isItExist = emailListData?.invitations.some(
-      (invitation: any) => invitation.invitee.email === inviteEmailInput,
-    );
-    if (canIInvite) {
-      if (!isItExist) {
-        await postData(`dashboards/${id}/invitations`, {
-          email: inviteEmailInput,
-        });
-        const emailListResponse = await getData(`dashboards/${id}/invitations`);
-        setEmailListData(emailListResponse.data);
-        setInviteModalState(false);
-      } else {
-        await setErrorModal(true);
-        setInviteModalState(false);
-      }
-    }
-  };
-
   useEffect(() => {
     const fetchData = async () => {
       try {
@@ -160,13 +118,6 @@ export default function Edit() {
 
     fetchData();
   }, [id]);
-
-  useEffect(() => {
-    //이메일인지 검사하고 맞으면 true값 반환
-    const reg =
-      /^[0-9a-zA-Z]([-_\.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_\.]?[0-9a-zA-Z])*\.[a-zA-Z]{2,3}$/i;
-    setCanIInvite(reg.test(inviteEmailInput));
-  }, [inviteEmailInput]);
 
   return (
     <S.Background>
@@ -211,45 +162,6 @@ export default function Edit() {
           </S.DashboardSettings>
         </S.MainContainer>
       </S.DashboardContainer>
-
-      {inviteModalState ? (
-        <S.ModalContainer onClick={handleSetInviteModalStateFalse}>
-          <Modal
-            title="초대하기"
-            name="이메일"
-            Placeholder="이메일"
-            cancelButton="취소"
-            submitButton="초대"
-            cancel={handleSetInviteModalStateFalse}
-            submit={handleInviteEmail}
-            value={setInviteEmailInput}
-          />
-        </S.ModalContainer>
-      ) : (
-        ""
-      )}
-      {errorModal ? (
-        <S.ModalContainer onClick={handleSetInviteModalStateFalse}>
-          <ModalCheckIt
-            text={"이미 초대된 회원입니다."}
-            submitButton={"확인"}
-            wrong={handleSetInviteModalStateFalse}
-          />
-        </S.ModalContainer>
-      ) : (
-        ""
-      )}
-      {errorModal2 ? (
-        <S.ModalContainer onClick={handleSetInviteModalStateFalse}>
-          <ModalCheckIt
-            text={"존재하지 않는 회원입니다."}
-            submitButton={"확인"}
-            wrong={handleSetInviteModalStateFalse}
-          />
-        </S.ModalContainer>
-      ) : (
-        ""
-      )}
     </S.Background>
   );
 }
